ちゃんと数字を抽出する

+++ web/models.py
@@ -4,7 +4,7 @@ from django.db import models
 class Rin(models.Model):
     id = models.AutoField(primary_key=True)
     tree_species = models.CharField(max_length=255, null=False)
-    diameter = models.IntegerField(null=False)
+    diameter = models.FloatField(null=False)
     latitude = models.FloatField(null=False)
     longitude = models.FloatField(null=False)
     created_at = models.DateTimeField(auto_now_add=True)
